+++ src/App.js
@@ -5,10 +5,7 @@ import MUIDataTable from 'mui-datatables';
 import TextFile from '@mui/material/TextField';
 import './App.css';
 import Autocomplete from '@mui/material/Autocomplete';
-import Modal from '@mui/material/Modal';
-import Box from '@mui/material/Box';
-import Typography from '@mui/material/Typography';
-
+import PokemonModal from './components/modal';
 
 function App() {
   const [data, setData] = useState([])
@@ -18,10 +15,7 @@ function App() {
   const handleClose = () => {
     setOpen(false)
   }
-
-
-
-
+  
   const columns = [
     {
       name: 'name',
@@ -32,17 +26,7 @@ function App() {
       label: 'url'
     }
   ]
-  const style = {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: 400,
-    bgcolor: 'background.paper',
-    border: '2px solid #000',
-    boxShadow: 24,
-    p: 4,
-  };
+
   const names = data.map(pokemon => {
     let properties = {
       label: pokemon.name,
@@ -72,21 +56,13 @@ function App() {
           sx={{ width: 300 }}
           renderInput={(params) => <TextFile {...params} label='Pokemones' />}
         />
-        <Modal
+        <PokemonModal
+          name={name}
           open={open}
-          onClose={handleClose}
-          aria-labelledby="modal-modal-title"
-          aria-describedby="modal-modal-description"
+          handleClose={handleClose}
         >
-          <Box sx={style}>
-            <Typography id="modal-modal-title" variant="h6" component="h2">
-              Pokemon
-            </Typography>
-            <Typography id="modal-modal-description" sx={{ mt: 2 }}>
-              {`${name} is the name of the pokemon you shoose`}
-            </Typography>
-          </Box>
-        </Modal>
+
+        </PokemonModal>
       </div>
       <div className="App">
         <MUIDataTable
