+++ src/components/modal.js
@@ -1,13 +1,47 @@
 import React from "react"
+import Modal from '@mui/material/Modal';
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
 
-const Modal = (name) => {
+const PokemonModal = (props) => {
 
-   
+ /*  const name = props.name
+  const open = props.open
+  const handleClose = props.handleClose */
+  // destructuring
+  const {name, open, handleClose} = props
 
-    return(
-        <>
-        
-        </>
-    )
+  const style = {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    transform: 'translate(-50%, -50%)',
+    width: 400,
+    bgcolor: 'background.paper',
+    border: '2px solid #000',
+    boxShadow: 24,
+    p: 4,
+  };
+
+  return (
+    <>
+      <Modal
+        open={open}
+        onClose={handleClose}
+        aria-labelledby="modal-modal-title"
+        aria-describedby="modal-modal-description"
+      >
+        <Box sx={style}>
+          <Typography id="modal-modal-title" variant="h6" component="h2">
+            Pokemon
+          </Typography>
+          <Typography id="modal-modal-description" sx={{ mt: 2 }}>
+            {`${name} is the name of the pokemon you choose`}
+          </Typography>
+        </Box>
+      </Modal>
+    </>
+  )
 }
-export default Modal
+
+export default PokemonModal
